Update cli.py
editing the repository to import json with my GeoJSONFarm2 file. I changed all json.dumps to json.dump because the file is that dump() takes two positional arguments: (1) the data object to be serialized, and (2) the file-like object to which the bytes will be written.--will this cause an error?